#10 ApplicationScope moved to ru.ltst.u2020mvp
