+++ app/src/internalDebug/java/ru/ltst/u2020mvp/ui/DebugUiModule.java
@@ -2,6 +2,7 @@ package ru.ltst.u2020mvp.ui;
 
 import dagger.Module;
 import dagger.Provides;
+import ru.ltst.u2020mvp.ApplicationScope;
 import ru.ltst.u2020mvp.base.mvp.Registry;
 import ru.ltst.u2020mvp.ui.debug.DebugAppContainer;
 import ru.ltst.u2020mvp.ui.annotation.ActivityScreenSwitcherServer;
