+++ build.gradle
@@ -3,7 +3,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.3.0-beta4'
+        classpath 'com.android.tools.build:gradle:1.3.0'
     }
 }
